Saved downloaded image to CD and displayed it in favorite cell.
