+++ NewsApp/ViewModel/FavoriteArticlesViewModel.swift
@@ -26,24 +26,4 @@ class FavoriteArticlesViewModel {
         self.articles = articles
         delegate?.articlesLoaded(articles: articles)
     }
-    
-    func addArticle(article: Article) {
-        coreDataManager.addArticle(article: article)
-        loadArticles()
-    }
-    
-    func getImage(urlString: String, completion: @escaping (UIImage?) -> ()) {
-//        if let coreDataManager.getArticleEntity(with: <#T##String#>)
-    }
-    
-    func downloadImage(urlString: String, completion: @escaping (UIImage?) -> ()) {
-        articleService.loadArticleImage(urlString: urlString) { image in
-            if let image, let imageData = image.pngData() {
-                self.coreDataManager.saveImage(for: urlString, with: imageData)
-                completion(image)
-            } else {
-                print("Couldn't load image in FavoriteArticlesViewModel")
-            }
-        }
-    }
 }
